Split core.ts into per-command and shared modules
core.ts had grown to mix diff parsing, coverage running, formatting,
and typecheck logic into one file. Move each concern to a dedicated
module so commands and shared utilities are easy to find:

- src/shared/{config,diff,coverage,format,github}.ts: cross-command
  utilities consumed by measure/diff/review/typecheck/mcp
- src/commands/typecheck/{typecheck,format}.ts: typecheck-specific
  logic and formatter
- src/commands/review/review.ts: moved from src/review.ts to live next
  to its CLI registration

Tests follow their source modules. Also retire the dead runJestCoverage
alias and update package.json main now that core.ts is gone.
