+++ src/runner/vitest-runner.test.ts
@@ -8,7 +8,8 @@ vi.mock("execa", () => ({
 }));
 
 import { execa } from "execa";
-import type { DiffFile, RunOptions } from "../core.js";
+import type { RunOptions } from "../shared/coverage.js";
+import type { DiffFile } from "../shared/diff.js";
 import { runVitest } from "./vitest.js";
 
 const mockExeca = vi.mocked(execa);
